Split providers into their own modules (#47)
* Added tiingo provider

* Split providers

* Removed old provider code

* Refactored lambda structure

* Fix packages
