+++ packages/salmon-lambda-functions/package.json
@@ -33,7 +33,7 @@
   ],
   "scripts": {
     "build": "tsc",
-    "pack": "ncc build ./src/index.ts --source-map"
+    "pack": "for x in ./src/*; do ncc build $x --source-map -o dist/$(basename $x .ts); done"
   },
   "peerDependencies": {
     "bignumber.js": "^9.0.1"
@@ -43,7 +43,9 @@
     "@defichain/jellyfish-transaction": "next",
     "@defichain/jellyfish-transaction-builder": "next",
     "@defichain/salmon-price-functions": "0.0.0",
-    "@defichain/salmon-oracles-functions": "0.0.0"
+    "@defichain/salmon-provider-tiingo": "0.0.0",
+    "@defichain/salmon-provider-finnhubb": "0.0.0",
+    "@defichain/salmon-provider-iexcloud": "0.0.0"
   },
   "devDependencies": {
     "@defichain/jellyfish-api-jsonrpc": "next",
