+++ packages/salmon-lambda-functions/src/tiingo.ts
@@ -1,8 +1,9 @@
-import { PriceManager, PriceSourceConfig, TiingoPriceProvider } from '@defichain/salmon-price-functions'
+import { PriceManager, PriceSourceConfig } from '@defichain/salmon-price-functions'
+import { TiingoPriceProvider } from '@defichain/salmon-provider-tiingo'
 
 export async function handler (event?: any): Promise<any> {
   const tiingoConfig: PriceSourceConfig = {
-    symbols: ['TSLA']
+    symbols: (process.env.SYMBOLS ?? '').split(',')
   }
 
   const priceManager = new PriceManager(tiingoConfig, new TiingoPriceProvider(process.env.API_TOKEN ?? ''))
