+++ packages/salmon-price-functions/src/index.ts
@@ -4,4 +4,3 @@
 
 export * from './price_manager'
 export * from './price_provider'
-export * from './providers'
