+++ packages/salmon-provider-finnhubb/README.md
@@ -1,0 +1,5 @@
+# @defichain/salmon-provider-finnhubb
+
+### Config
+
+Requires a finnhubb API token
