+++ packages/salmon-provider-tiingo/README.md
@@ -1,0 +1,5 @@
+# @defichain/salmon-provider-tiingo
+
+### Config
+
+Requires a Tiingo API token
