+++ tsconfig.json
@@ -5,6 +5,9 @@
     "paths": {
       "@defichain/salmon-oracles-functions": ["salmon-oracles-functions/src"],
       "@defichain/salmon-price-functions": ["salmon-price-functions/src"],
+      "@defichain/salmon-provider-tiingo": ["salmon-provider-tiingo/src"],
+      "@defichain/salmon-provider-iexcloud": ["salmon-provider-iexcloud/src"],
+      "@defichain/salmon-provider-finnhubb": ["salmon-provider-finnhubb/src"],
       "@defichain/salmon-lambda-functions": ["salmon-lambda-functions/src"]
     }
   }
